feat(frontend): the secret and vault tables move into @agenta/settings-ui
Both were already package-backed on the data side (useVaultSecret from
@agenta/entities/secret), so what remained was de-antd'ing them and lifting the
dialogs out. The model-registry dialogs and the configure-secret modal become
slots: they belong to the registry, not to settings, and each host renders its
own.

dateTimeHelper moves to @agenta/shared/utils/dateTime on the way — both tables
need formatDay, and it has 21 consumers across oss and ee, which all keep their
import path through a re-export.
